% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_analysis.R
\name{tol_common_genes}
\alias{tol_common_genes}
\title{Choose a gene shared by multiple downloaded FASTA files}
\usage{
tol_common_genes(plan, min_records = 3)
}
\arguments{
\item{plan}{A data frame returned by \code{\link[=tol_attach_fasta]{tol_attach_fasta()}} or a FASTA plan
with local files.}

\item{min_records}{Minimum number of records that must contain the gene.}
}
\value{
A tibble ranked by number of records and median width.
}
\description{
Choose a gene shared by multiple downloaded FASTA files
}
