% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{tol_data_dir}
\alias{tol_data_dir}
\title{Local data directory for downloaded files}
\usage{
tol_data_dir(create = TRUE)
}
\arguments{
\item{create}{Create the directory if it does not exist.}
}
\value{
A normalized path.
}
\description{
The default can be overridden with option \code{rtreeoflife.data_dir} or
environment variable \code{RTREEOFLIFE_DATA_DIR}.
}
