% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_analysis.R
\name{tol_fasta_long}
\alias{tol_fasta_long}
\title{Convert attached FASTA data to a tidy long table}
\usage{
tol_fasta_long(plan, fasta_column = "fasta")
}
\arguments{
\item{plan}{A data frame returned by \code{\link[=tol_attach_fasta]{tol_attach_fasta()}} or a FASTA plan
with local files.}

\item{fasta_column}{Name of the FASTA list-column. If it is not present,
\code{\link[=tol_attach_fasta]{tol_attach_fasta()}} is called.}
}
\value{
A tibble with one row per sequence.
}
\description{
Convert attached FASTA data to a tidy long table
}
