% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{tol_known_bundle}
\alias{tol_known_bundle}
\title{Known download bundles}
\usage{
tol_known_bundle()
}
\value{
A data frame with bundle names and included remote paths.
}
\description{
Alias for \code{\link[=tol_known_bundles]{tol_known_bundles()}}.
}
