% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{tol_known_bundles}
\alias{tol_known_bundles}
\title{Known download bundles}
\usage{
tol_known_bundles()
}
\value{
A data frame with bundle names and included remote paths.
}
\description{
Known download bundles
}
