% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{tol_manifest}
\alias{tol_manifest}
\title{Read a downloaded Kew Tree of Life manifest}
\usage{
tol_manifest(file, manifest = NULL, ...)
}
\arguments{
\item{file}{Local path to a manifest file.}

\item{manifest}{Manifest type. Used only to assign column names when known.}

\item{...}{Additional arguments passed to \code{\link[utils:read.delim]{utils::read.delim()}}.}
}
\value{
A data frame.
}
\description{
Read a downloaded Kew Tree of Life manifest
}
