% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_analysis.R
\name{tol_plot_fasta_summary}
\alias{tol_plot_fasta_summary}
\title{Plot FASTA summary with ggplot2}
\usage{
tol_plot_fasta_summary(summary)
}
\arguments{
\item{summary}{A summary returned by \code{\link[=tol_fasta_summary]{tol_fasta_summary()}}.}
}
\value{
A ggplot object.
}
\description{
Plot FASTA summary with ggplot2
}
