% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_analysis.R
\name{tol_plot_gene_recovery}
\alias{tol_plot_gene_recovery}
\title{Plot recovered gene counts with ggplot2}
\usage{
tol_plot_gene_recovery(records)
}
\arguments{
\item{records}{Species records returned by \code{\link[=tol_species_index]{tol_species_index()}} or
\code{\link[=tol_search_species]{tol_search_species()}}.}
}
\value{
A ggplot object.
}
\description{
Plot recovered gene counts with ggplot2
}
