% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{tol_search_species}
\alias{tol_search_species}
\title{Search Kew Tree of Life species records}
\usage{
tol_search_species(
  index = tol_species_index(),
  sequence_id = NULL,
  order = NULL,
  family = NULL,
  genus = NULL,
  specific_epithet = NULL,
  scientific_name = NULL,
  query = NULL,
  ignore_case = TRUE
)
}
\arguments{
\item{index}{A species index returned by \code{\link[=tol_species_index]{tol_species_index()}}. If omitted,
the built-in \link{tol_species} dataset is used.}

\item{sequence_id}{Optional sequence identifier or vector of identifiers.}

\item{order, family, genus, specific_epithet, scientific_name}{Optional taxonomic
filters.}

\item{query}{Optional free-text query.}

\item{ignore_case}{Ignore case in text filters.}
}
\value{
A filtered data frame.
}
\description{
Search Kew Tree of Life species records
}
