% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{tol_set_base_url}
\alias{tol_set_base_url}
\title{Set the Kew Tree of Life base URL for the current R session}
\usage{
tol_set_base_url(base_url = "https://sftp.kew.org/pub/treeoflife")
}
\arguments{
\item{base_url}{Base URL without the release directory.}
}
\value{
The previous value of option \code{rtreeoflife.base_url}, invisibly.
}
\description{
Set the Kew Tree of Life base URL for the current R session
}
