% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{tol_species_index}
\alias{tol_species_index}
\title{Read the Kew Tree of Life species index}
\usage{
tol_species_index(file = NULL, use_package_data = TRUE)
}
\arguments{
\item{file}{Local path to \code{all_species_list.csv}. If \code{NULL}, the package
dataset \link{tol_species} is returned.}

\item{use_package_data}{Use the built-in \link{tol_species} dataset when
\code{file = NULL}.}
}
\value{
A data frame with normalized column names and derived fields.
}
\description{
Read the Kew Tree of Life species index
}
